Restored multi-segment reader for CRX
